requirements.txt
python-telegram-bot>=13.0
requests>=2.25.1